PiggyBank/requirements.txt
python-jose~=3.5.0
python-dotenv~=1.2.1
fastapi~=0.124.4
SQLAlchemy~=2.0.45
security~=1.3.1
pydantic~=2.12.5
passlib[bcrypt]
bcrypt
python-multipart
python-jose[cryptography]
pydantic[email]
pytest~=9.0.2
psycopg2
uvicorn